캐릭터 사망에 대한 처리

+++ LittleSword/Assets/02_Scripts/Enemy/Stats/EnemyStats.cs
@@ -13,7 +13,7 @@ public class EnemyStats : ScriptableObject
     [Header("Combat Stats")]
     public float chaseDistance = 5f;
     public float attackDistance = 1.5f;
-    public float attackDamage = 10f;
+    public int attackDamage = 10;
     public float attackCooldown = 1f;
 
 }
